[Backend] handle invalid id on user controller

+++ Backend/controllers/userController.js
@@ -1,12 +1,19 @@
 const User = require('./../models/userModel');
+const {
+  Types: { ObjectId },
+} = require('mongoose');
+
+const isIdValid = id => {
+  return ObjectId.isValid(id) && new ObjectId(id).toString === id;
+};
 
 exports.createUser = async (req, res) => {
   try {
     const user = await User.create(req.body);
 
     res.status(201).json({
-      status: 'ok',
-      data: user,
+      status: 'success',
+      data: { user },
     });
   } catch (e) {
     throw err;
@@ -16,15 +23,18 @@ exports.getUsers = async (req, res) => {
   try {
     const users = await User.find();
     res.status(200).json({
-      status: 'ok',
+      status: 'success',
       results: users.length,
-      data: users,
+      data: { users },
     });
   } catch (err) {
     throw err;
   }
 };
 exports.getUser = async (req, res) => {
+  if (!isIdValid(req.params.id)) {
+    throw err;
+  }
   try {
     const user = await User.findById(req.params.id);
     res.status(200).json({
@@ -36,14 +46,17 @@ exports.getUser = async (req, res) => {
   }
 };
 exports.updateUser = async (req, res) => {
-  console.log('updateing');
+  //   console.log('updateing');
+  if (!isIdValid(req.params.id)) {
+    throw err;
+  }
   try {
     const newUser = await User.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
     });
     res.status(200).json({
-      status: 'ok',
-      data: newUser,
+      status: 'sucess',
+      data: { newUser },
     });
   } catch (err) {
     // console.log(err.message);
@@ -51,11 +64,14 @@ exports.updateUser = async (req, res) => {
   }
 };
 exports.deleteUser = async (req, res) => {
+  if (!isIdValid(req.params.id)) {
+    throw err;
+  }
   try {
     const user = await User.findByIdAndDelete(req.params.id);
     res.status(200).json({
-      staus: 'ok',
-      data: user,
+      staus: 'success',
+      data: { user },
     });
   } catch (err) {
     throw err;
